+++ changelog.md
@@ -1,2 +1,9 @@
+<h3>[v1.0.1]</h3>
+- Added BlockScanner class
+- Added UnitMessenger class
+---
+<h3>[v1.0.1]</h3>
+- Removed useless registry classes
+---
 <h3>[v1.0.0]</h3>
 - Initial build.